requirements.txt
# Runtime dependencies - These are also specified in pyproject.toml
# This file is for convenience during development
pandas>=1.3.0
requests>=2.27.1
tabulate>=0.9.0
python-telegram-bot>=20.0
# For hosting, we also need the local package
.

# Development dependencies
# To install with dev dependencies, use: pip install -e ".[dev]"
